requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
tqdm>=4.62.0
tensorboard>=2.7.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
scipy>=1.7.0
albumentations>=1.3.0
pyyaml>=6.0
segmentation-models-pytorch>=0.3.0
